native: the angle buffer was 4-byte strided - the view rotation is fixed
The axis permutation the last commit measured (world X arriving as
camera Y) was the 16-bit width trap again, in scenegraph.c. Every
NodeApplyTransform_* writes the three BAM angles into the scratch at
0xab5208, and the original stores them as WORDS - mov word ptr
[0xab5208] / [0xab520a] / [0xab520c] - while every BuildRotMatrix_Order*
reads them back as short[3]. The twin declared the scratch as
unsigned int *, so [0]/[1]/[2] wrote at a 4-byte stride: the yaw landed
in the roll slot, the roll fell past the array, and the pitch slot kept
whatever was there. One character of type, and the view matrix came out
as an axis permutation.

Measured before and after, same frame, MK4_ARENA_INTRO=1, eye at
(-7.0, -1.5, 0.0) with the geometry at the origin:

  before   camera-space (0, -7.0, 1.5)   world X arriving as camera Y
  after    camera-space (-1.5, 0, 7.0)   7.0 units of real DEPTH

7.0 is exactly the eye-to-origin distance, so the transform is now
doing what it should. With the intro armed the fight frame stops being
black during the opening: a compact object appears at proper
perspective scale, proportioned like a figure, from frame 172 to about
frame 210 (~1800 pixels), then the camera spins past it.

The default build is unchanged (19972 px), because without the intro
the camera-space depth is still zero and the projection still divides
by ~nothing - that is the other half of the framing problem and it is
untouched here.

Gates: matching BYTE-IDENTICAL; native-frame-check 61800px;
native-char-check byte-identical; moves screen 307200/307200.

+++ src/engine/scenegraph.c
@@ -20,7 +20,12 @@
 #define g_xformDirtyFlags (*(unsigned int *)MK4_VA(unsigned int, 0x54208cu))
 #define g_xformEntityIdx (*(unsigned int *)MK4_VA(unsigned int, 0x542048u))
 #define g_xformLoopCounter (*(unsigned int *)MK4_VA(unsigned int, 0x53a1acu))
-#define g_xformTempAngles ((unsigned int *)MK4_VA(unsigned int, 0xab5208u))
+/* The original stores the three BAM angles as WORDS at 0xab5208 /
+ * 0xab520a / 0xab520c (mov word ptr), and every BuildRotMatrix_Order*
+ * reads them as short[3]. Typed as unsigned int * the stride was 4, so
+ * the yaw landed in the roll slot and the view rotation came out as an
+ * axis permutation - see [[feedback_width16_matrix_trap]], same family. */
+#define g_xformTempAngles ((short *)MK4_VA(short, 0xab5208u))
 #endif
